+++ trackme/bin/trackme_rest_handler_smart_status.py
@@ -111,8 +111,10 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                 + "| `trackme_lookup_data_sampling`"\
                 + "| nomv tags"\
                 + "| eval data_last_lag_seen=round(data_last_lag_seen, 0), data_last_time_seen=round(data_last_time_seen, 0)"\
+                + "| `trackme_eval_icons`"\
+                + "| `trackme_smart_status_weekdays_to_human`"\
                 + "| fillnull value=\"null\" data_name data_index data_sourcetype data_source_state data_lag_alert_kpis"\
-                + "data_last_lag_seen data_last_ingestion_lag_seen data_max_lag_allowed data_last_time_seen enable_behaviour_analytic dcount_host "\
+                + "data_last_lag_seen data_last_ingestion_lag_seen data_max_lag_allowed data_last_time_seen data_monitoring_wdays status_message enable_behaviour_analytic dcount_host "\
                 + "min_dcount_host isOutlier isAnomaly data_sample_anomaly_reason data_sample_status_colour data_sample_status_message "\
                 + "elastic_source_from_part1 elastic_source_from_part2 elastic_source_search_constraint elastic_source_search_mode"
 
@@ -133,6 +135,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                     data_last_ingestion_lag_seen = query_result["data_last_ingestion_lag_seen"]
                     data_max_lag_allowed = query_result["data_max_lag_allowed"]
                     data_last_time_seen = query_result["data_last_time_seen"]
+                    data_monitoring_wdays = query_result["data_monitoring_wdays"]
                     enable_behaviour_analytic = query_result["enable_behaviour_analytic"]
                     dcount_host = query_result["dcount_host"]
                     min_dcount_host = query_result["min_dcount_host"]
@@ -145,6 +148,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                     elastic_source_search_mode = query_result["elastic_source_search_mode"]
                     elastic_source_from_part1 = query_result["elastic_source_from_part1"]
                     elastic_source_from_part2 = query_result["elastic_source_from_part2"]
+                    status_message = query_result["status_message"]
 
                 except Exception as e:
                     data_index = None
@@ -155,6 +159,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                     data_last_ingestion_lag_seen = None
                     data_max_lag_allowed = None
                     data_last_time_seen = None
+                    data_monitoring_wdays = None
                     enable_behaviour_analytic = None
                     dcount_host = None
                     min_dcount_host = None
@@ -167,6 +172,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                     elastic_source_search_mode = None
                     elastic_source_from_part1 = None
                     elastic_source_from_part2 = None
+                    status_message = None
 
                 # min_dcount_host defaults to str any, for ease of code, comvert to 0
                 if min_dcount_host in ("any"):
@@ -263,7 +269,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                     # subcase: data is in the future
                     ###############################################################
 
-                    if data_source_state in ("orange"):
+                    if data_source_state in ("orange") and not (re.search("week days rules conditions are not met", str(status_message)) or isOutlier in ("1")):
 
                         # Get lagging statistics from live data
 
@@ -810,10 +816,16 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                         # increment the smart_code by 10
                         smart_code += 10
 
+                        if data_source_state in ("orange") and re.search("week days rules conditions are not met", str(status_message)):
+                            smart_result = "TrackMe triggered an alert due to the latest data available that is out of the acceptable window, the maximal event lag allowed is: " + str(data_max_lag_allowed) + " seconds, while the latest data available is: " + str(human_last_datetime) + ", the data is late by: " + str(current_delay) + " (days, HH:MM:SS)"\
+                                + ", however due to week days monitoring rules (" + str(data_monitoring_wdays) + "), the entity is currently in a state that will not generate an active alert."
+                        else:
+                            smart_result = "TrackMe triggered an alert due to the latest data available that is out of the acceptable window, the maximal event lag allowed is: " + str(data_max_lag_allowed) + " seconds, while the latest data available is: " + str(human_last_datetime) + ", the data is late by: " + str(current_delay) + " (days, HH:MM:SS)"
+
                         results_message = '{' \
-                        + '"data_name": "' + str(data_name)  + '", '\
+                        + '"data_name": "' + str(data_name) + '", '\
                         + '"data_source_state": "' + str(data_source_state)  + '", '\
-                        + '"smart_result": "TrackMe triggered an alert due to the latest data available that is out of the acceptable window, the maximal event lag allowed is: ' + str(data_max_lag_allowed) + ' seconds, while the latest data available is: ' + str(human_last_datetime) + ', the data is late by: ' + str(current_delay) + ' (days, HH:MM:SS)", '\
+                        + '"smart_result": "' + str(smart_result) + '", '\
                         + '"' + str(report_name) + '": "' + str(report_desc) + str(summary) + '", '\
                         + '"smart_code": "' + str(smart_code) + '", ' \
                         + '"correlation_flipping_state": "' + str(flipping_correlation_msg) + '", '\
@@ -1057,10 +1069,16 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                         # increment the smart_code by 20
                         smart_code += 20
 
+                        if data_source_state in ("orange") and re.search("week days rules conditions are not met", str(status_message)):
+                            smart_result = "TrackMe triggered an alert due to indexing lag detected out of the acceptable window, the maximal ingestion lag allowed is: " + str(data_max_lag_allowed) + " seconds, while an ingestion lag of " + str(datetime.timedelta(seconds=int(data_last_ingestion_lag_seen))) + " (days, HH:MM:SS) was detected, review the hosts attached to this report to investigate the root cause."\
+                                + " However due to week days monitoring rules (" + str(data_monitoring_wdays) + "), the entity is currently in a state that will not generate an active alert."
+                        else:
+                            smart_result = "TrackMe triggered an alert due to indexing lag detected out of the acceptable window, the maximal ingestion lag allowed is: " + str(data_max_lag_allowed) + " seconds, while an ingestion lag of " + str(datetime.timedelta(seconds=int(data_last_ingestion_lag_seen))) + " (days, HH:MM:SS) was detected, review the hosts attached to this report to investigate the root cause."
+
                         results_message = '{' \
                         + '"data_name": "' + str(data_name) + '", '\
                         + '"data_source_state": "' + str(data_source_state) + '", '\
-                        + '"smart_result": "TrackMe triggered an alert due to indexing lag detected out of the acceptable window, the maximal ingestion lag allowed is: ' + str(data_max_lag_allowed) + ' seconds, while an ingestion lag of ' + str(datetime.timedelta(seconds=int(data_last_ingestion_lag_seen))) + ' (days, HH:MM:SS) was detected, review the hosts attached to this report to investigate the root cause.", '\
+                        + '"smart_result": "' + str(smart_result) + '", '\
                         + '"' + str(report_name) + '": "' + str(report_desc) + str(summary) + '", '\
                         + '"smart_code": "' + str(smart_code) + '", ' \
                         + '"correlation_flipping_state": "' + str(flipping_correlation_msg) + '", '\
@@ -1311,9 +1329,15 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                             summary = None
 
                         # define the message
-                        smart_result_msg = 'TrackMe triggered an alert on ' + str(flipping_time) + ' due to the minimal distinct count of hosts configured for this data source (threshold: '\
-                            + str(min_dcount_host) + ' hosts) which condition is not met as only ' + str(dcount_host) \
-                            + ' hosts are detected currently. Review this threshold and the current data source activity accordingly.'
+                        if data_source_state in ("orange") and re.search("week days rules conditions are not met", str(status_message)):
+                            smart_result = "TrackMe triggered an alert on " + str(flipping_time) + " due to the minimal distinct count of hosts configured for this data source (threshold: "\
+                            + str(min_dcount_host) + " hosts) which condition is not met as only " + str(dcount_host) \
+                            + " hosts are detected currently. Review this threshold and the current data source activity accordingly."\
+                            + " However due to week days monitoring rules (" + str(data_monitoring_wdays) + "), the entity is currently in a state that will not generate an active alert."
+                        else:
+                            smart_result = "TrackMe triggered an alert on " + str(flipping_time) + " due to the minimal distinct count of hosts configured for this data source (threshold: "\
+                            + str(min_dcount_host) + " hosts) which condition is not met as only " + str(dcount_host)\
+                            + " hosts are detected currently. Review this threshold and the current data source activity accordingly."
 
                         # increment the smart_code by 30
                         smart_code += 30
@@ -1321,7 +1345,7 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                         results = '{' \
                         + '"data_name": "' + data_name  + '", '\
                         + '"data_source_state": "' + data_source_state  + '", '\
-                        + '"smart_result": "' + str(smart_result_msg) + '", '\
+                        + '"smart_result": "' + str(smart_result) + '", '\
                         + '"' + str(report_name) + '": "' + str(report_desc) + str(summary) + '", '\
                         + '"smart_code": "' + str(smart_code) + '", ' \
                         + '"correlation_flipping_state": "' + str(flipping_correlation_msg) + '", '\
@@ -1409,14 +1433,25 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                         # increment the smart_code by 40
                         smart_code += 40
 
+                        # define the message
+                        if data_source_state in ("orange"):
+                            smart_result = "TrackMe triggered an alert on this data source due to outliers detection in the "\
+                            + "event count, outliers are based on the calculation of a lower and upper bound (if alerting on upper) determined "\
+                            + "against the data source usual behaviour and outliers parameters. Review the correlation results to determine "\
+                            + "if the behaviour is expected or symptomatic of an issue happening on the data source (lost of "\
+                            + "sources or hosts, etc.) and proceed to any outliers configuration fine tuning if necessary."\
+                            + " However due to week days monitoring rules (" + str(data_monitoring_wdays) + "), the entity is currently in a state that will not generate an active alert."
+                        else:
+                            smart_result =  "TrackMe triggered an alert on this data source due to outliers detection in the "\
+                            + "event count, outliers are based on the calculation of a lower and upper bound (if alerting on upper) determined "\
+                            + "against the data source usual behaviour and outliers parameters. Review the correlation results to determine "\
+                            + "if the behaviour is expected or symptomatic of an issue happening on the data source (lost of "\
+                            + "sources or hosts, etc.) and proceed to any outliers configuration fine tuning if necessary."\
+
                         results = '{' \
                         + '"data_name": "' + data_name  + '", '\
-                        + '"data_source_state": "' + data_source_state  + '", '\
-                        + '"smart_result": "TrackMe triggered an alert on this data source due to outliers detection in the '\
-                        + 'event count, outliers are based on the calculation of a lower and upper bound (if alerting on upper) determined '\
-                        + 'against the data source usual behaviour and outliers parameters. Review the correlation results to determine '\
-                        + 'if the behaviour is expected or symptomatic of an issue happening on the data source (lost of '\
-                        + 'sources or hosts, etc.) and proceed to any outliers configuration fine tuning if necessary.", '\
+                        + '"data_source_state": "' + data_source_state + '", '\
+                        + '"smart_result": "' + str(smart_result) + '", '\
                         + '"smart_code": "' + str(smart_code) + '", ' \
                         + '"correlation_outliers": "[ description: Last 24h outliers detection ], [ OutliersCount: ' \
                         + str(countOutliers) + ' ], [ latest4hcount: ' + str(latest4hcount) + ' ], [ lowerBound: ' \
@@ -1674,11 +1709,20 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
 
                         # Result
 
+                        # define the message
+                        if data_source_state in ("orange") and re.search("week days rules conditions are not met", str(status_message)):
+                            smart_result = "TrackMe triggered an alert due to anomaly detection in the data sampling worfklow (reason: " + str(anomaly_reason)\
+                            + " detected on " + str(anomaly_mtime) + "), "\
+                            + " However due to week days monitoring rules (" + str(data_monitoring_wdays) + "), the entity is currently in a state that will not generate an active alert."\
+                            + ", [ message: " + str(smart_correlation) + " ]"
+                        else:
+                            smart_result =  "TrackMe triggered an alert due to anomaly detection in the data sampling worfklow (reason: " + str(anomaly_reason)\
+                            + " detected on " + str(anomaly_mtime) + "), [ message: " + str(smart_correlation) + " ]"
+
                         results = '{' \
                         + '"data_name": "' + data_name  + '", '\
-                        + '"data_source_state": "' + data_source_state  + '", '\
-                        + '"smart_result": "TrackMe triggered an alert due to anomaly detection in the data sampling worfklow (reason: ' + str(anomaly_reason) \
-                        + ' detected on ' + str(anomaly_mtime) + '), [ message: ' + str(smart_correlation) + ' ]", '\
+                        + '"data_source_state": "' + data_source_state + '", '\
+                        + '"smart_result": "' + str(smart_result) + '", '\
                         + '"smart_code": "' + str(smart_code) + '", ' \
                         + '"correlation_data_sampling": "description: [ Last 4h top event count/model ], ' + str(data_sampling_correlation) + '", ' \
                         + '"correlation_flipping_state": "' + str(flipping_correlation_msg) + '"'\
@@ -1702,6 +1746,8 @@ class TrackMeHandlerSmartStatus_v1(rest_handler.RESTHandler):
                         + '"data_last_lag_seen": "' + str(data_last_lag_seen) + '", '\
                         + '"data_max_lag_allowed": "' + str(data_max_lag_allowed) + '", '\
                         + '"data_last_ingestion_lag_seen": "' + str(data_last_ingestion_lag_seen) + '", '\
+                        + '"data_monitoring_wdays": "' + str(data_monitoring_wdays) + '", '\
+                        + '"status_message": "' + str(status_message) + '", '\
                         + '"dcount_host": "' + str(dcount_host) + '", '\
                         + '"min_dcount_host": "' + str(min_dcount_host) + '", '\
                         + '"isOutlier": "' + str(isOutlier) + '", '\